requirements.txt
Flask
Flask-SQLAlchemy
pymysql
requests
Werkzeug
Flask-Cors
Flask-RESTful
pika
cryptography
python-dotenv